rename `elgamal_program` to `zk_elgamal_proof_program`
