+++ zk-sdk/src/zk_elgamal_proof_program/errors.rs
