+++ zk-sdk/src/zk_elgamal_proof_program/proof_data/batched_range_proof/batched_range_proof_u64.rs
@@ -1,69 +1,61 @@
-//! The 256-bit batched range proof instruction.
+//! The 64-bit batched range proof instruction.
 
-use {
-    crate::{
-        elgamal_program::proof_data::{
-            batched_range_proof::BatchedRangeProofContext, ProofType, ZkProofData,
-        },
-        range_proof::pod::PodRangeProofU256,
-    },
-    bytemuck::{Pod, Zeroable},
-};
 #[cfg(not(target_os = "solana"))]
 use {
     crate::{
-        elgamal_program::{
-            errors::{ProofGenerationError, ProofVerificationError},
-            proof_data::batched_range_proof::{MAX_COMMITMENTS, MAX_SINGLE_BIT_LENGTH},
-        },
         encryption::pedersen::{PedersenCommitment, PedersenOpening},
         range_proof::RangeProof,
+        zk_elgamal_proof_program::{
+            errors::{ProofGenerationError, ProofVerificationError},
+            proof_data::batched_range_proof::MAX_COMMITMENTS,
+        },
     },
     std::convert::TryInto,
 };
-
-#[cfg(not(target_os = "solana"))]
-const BATCHED_RANGE_PROOF_U256_BIT_LENGTH: usize = 256;
+use {
+    crate::{
+        range_proof::pod::PodRangeProofU64,
+        zk_elgamal_proof_program::proof_data::{
+            batched_range_proof::BatchedRangeProofContext, ProofType, ZkProofData,
+        },
+    },
+    bytemuck::{Pod, Zeroable},
+};
 
 /// The instruction data that is needed for the
-/// `ProofInstruction::BatchedRangeProofU256Data` instruction.
+/// `ProofInstruction::VerifyBatchedRangeProofU64` instruction.
 ///
 /// It includes the cryptographic proof as well as the context data information needed to verify
 /// the proof.
 #[derive(Clone, Copy, Pod, Zeroable)]
 #[repr(C)]
-pub struct BatchedRangeProofU256Data {
+pub struct BatchedRangeProofU64Data {
     /// The context data for a batched range proof
     pub context: BatchedRangeProofContext,
 
     /// The batched range proof
-    pub proof: PodRangeProofU256,
+    pub proof: PodRangeProofU64,
 }
 
 #[cfg(not(target_os = "solana"))]
-impl BatchedRangeProofU256Data {
+impl BatchedRangeProofU64Data {
     pub fn new(
         commitments: Vec<&PedersenCommitment>,
         amounts: Vec<u64>,
         bit_lengths: Vec<usize>,
         openings: Vec<&PedersenOpening>,
     ) -> Result<Self, ProofGenerationError> {
-        // Range proof on 256 bit length could potentially result in an unexpected behavior and
-        // therefore, restrict the bit length to be at most 128. This check is not needed for the
-        // `BatchedRangeProofU64` or `BatchedRangeProofU128`.
-        if bit_lengths
-            .iter()
-            .any(|length| *length > MAX_SINGLE_BIT_LENGTH)
-        {
-            return Err(ProofGenerationError::IllegalCommitmentLength);
-        }
-
-        // the sum of the bit lengths must be 256
+        // the sum of the bit lengths must be 64
         let batched_bit_length = bit_lengths
             .iter()
             .try_fold(0_usize, |acc, &x| acc.checked_add(x))
             .ok_or(ProofGenerationError::IllegalAmountBitLength)?;
-        if batched_bit_length != BATCHED_RANGE_PROOF_U256_BIT_LENGTH {
+
+        // `u64::BITS` is 64, which fits in a single byte and should not overflow to `usize` for an
+        // overwhelming number of platforms. However, to be extra cautious, use `try_from` and
+        // `unwrap` here. A simple case `u64::BITS as usize` can silently overflow.
+        let expected_bit_length = usize::try_from(u64::BITS).unwrap();
+        if batched_bit_length != expected_bit_length {
             return Err(ProofGenerationError::IllegalAmountBitLength);
         }
 
@@ -79,8 +71,8 @@ impl BatchedRangeProofU256Data {
     }
 }
 
-impl ZkProofData<BatchedRangeProofContext> for BatchedRangeProofU256Data {
-    const PROOF_TYPE: ProofType = ProofType::BatchedRangeProofU256;
+impl ZkProofData<BatchedRangeProofContext> for BatchedRangeProofU64Data {
+    const PROOF_TYPE: ProofType = ProofType::BatchedRangeProofU64;
 
     fn context_data(&self) -> &BatchedRangeProofContext {
         &self.context
@@ -90,13 +82,6 @@ impl ZkProofData<BatchedRangeProofContext> for BatchedRangeProofU256Data {
     fn verify_proof(&self) -> Result<(), ProofVerificationError> {
         let (commitments, bit_lengths) = self.context.try_into()?;
         let num_commitments = commitments.len();
-
-        if bit_lengths
-            .iter()
-            .any(|length| *length > MAX_SINGLE_BIT_LENGTH)
-        {
-            return Err(ProofVerificationError::IllegalCommitmentLength);
-        }
 
         if num_commitments > MAX_COMMITMENTS || num_commitments != bit_lengths.len() {
             return Err(ProofVerificationError::IllegalCommitmentLength);
@@ -116,14 +101,14 @@ mod test {
     use {
         super::*,
         crate::{
-            elgamal_program::errors::ProofVerificationError, encryption::pedersen::Pedersen,
-            range_proof::errors::RangeProofVerificationError,
+            encryption::pedersen::Pedersen, range_proof::errors::RangeProofVerificationError,
+            zk_elgamal_proof_program::errors::ProofVerificationError,
         },
     };
 
     #[test]
-    fn test_batched_range_proof_256_instruction_correctness() {
-        let amount_1 = 4294967295_u64;
+    fn test_batched_range_proof_u64_instruction_correctness() {
+        let amount_1 = 255_u64;
         let amount_2 = 77_u64;
         let amount_3 = 99_u64;
         let amount_4 = 99_u64;
@@ -141,7 +126,7 @@ mod test {
         let (commitment_7, opening_7) = Pedersen::new(amount_7);
         let (commitment_8, opening_8) = Pedersen::new(amount_8);
 
-        let proof_data = BatchedRangeProofU256Data::new(
+        let proof_data = BatchedRangeProofU64Data::new(
             vec![
                 &commitment_1,
                 &commitment_2,
@@ -155,7 +140,7 @@ mod test {
             vec![
                 amount_1, amount_2, amount_3, amount_4, amount_5, amount_6, amount_7, amount_8,
             ],
-            vec![32, 32, 32, 32, 32, 32, 32, 32],
+            vec![8, 8, 8, 8, 8, 8, 8, 8],
             vec![
                 &opening_1, &opening_2, &opening_3, &opening_4, &opening_5, &opening_6, &opening_7,
                 &opening_8,
@@ -165,7 +150,7 @@ mod test {
 
         assert!(proof_data.verify_proof().is_ok());
 
-        let amount_1 = 4294967296_u64; // not representable as a 32-bit number
+        let amount_1 = 256_u64; // not representable as an 8-bit number
         let amount_2 = 77_u64;
         let amount_3 = 99_u64;
         let amount_4 = 99_u64;
@@ -183,7 +168,7 @@ mod test {
         let (commitment_7, opening_7) = Pedersen::new(amount_7);
         let (commitment_8, opening_8) = Pedersen::new(amount_8);
 
-        let proof_data = BatchedRangeProofU256Data::new(
+        let proof_data = BatchedRangeProofU64Data::new(
             vec![
                 &commitment_1,
                 &commitment_2,
@@ -197,7 +182,7 @@ mod test {
             vec![
                 amount_1, amount_2, amount_3, amount_4, amount_5, amount_6, amount_7, amount_8,
             ],
-            vec![32, 32, 32, 32, 32, 32, 32, 32],
+            vec![8, 8, 8, 8, 8, 8, 8, 8],
             vec![
                 &opening_1, &opening_2, &opening_3, &opening_4, &opening_5, &opening_6, &opening_7,
                 &opening_8,
