+++ zk-sdk/src/zk_elgamal_proof_program/proof_data/errors.rs
