+++ zk-sdk/src/zk_elgamal_proof_program/proof_data/zero_ciphertext.rs
@@ -7,9 +7,9 @@
 #[cfg(not(target_os = "solana"))]
 use {
     crate::{
-        elgamal_program::errors::{ProofGenerationError, ProofVerificationError},
         encryption::elgamal::{ElGamalCiphertext, ElGamalKeypair},
         sigma_proofs::zero_ciphertext::ZeroCiphertextProof,
+        zk_elgamal_proof_program::errors::{ProofGenerationError, ProofVerificationError},
     },
     bytemuck::bytes_of,
     merlin::Transcript,
@@ -17,9 +17,9 @@ use {
 };
 use {
     crate::{
-        elgamal_program::proof_data::{ProofType, ZkProofData},
         encryption::pod::elgamal::{PodElGamalCiphertext, PodElGamalPubkey},
         sigma_proofs::pod::PodZeroCiphertextProof,
+        zk_elgamal_proof_program::proof_data::{ProofType, ZkProofData},
     },
     bytemuck::{Pod, Zeroable},
 };
