+++ zk-sdk/src/zk_elgamal_proof_program/state.rs
@@ -1,5 +1,5 @@
 use {
-    crate::elgamal_program::proof_data::{pod::PodProofType, ProofType},
+    crate::zk_elgamal_proof_program::proof_data::{pod::PodProofType, ProofType},
     bytemuck::{bytes_of, Pod, Zeroable},
     num_traits::ToPrimitive,
     solana_program::{
